+++ bpm-engine-impl/src/main/java/io/takari/bpm/elements/EventBasedGatewayHandler.java
@@ -5,7 +5,6 @@ import java.util.List;
 import io.takari.bpm.actions.Action;
 import io.takari.bpm.actions.FollowNewGroupAction;
 import io.takari.bpm.actions.PopCommandAction;
-import io.takari.bpm.actions.PushCommandAction;
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.commands.ProcessElementCommand;
 import io.takari.bpm.state.ProcessInstance;