+++ bpm-engine-impl/src/main/java/io/takari/bpm/elements/SubProcessHandler.java
@@ -1,19 +1,23 @@
 package io.takari.bpm.elements;
 
-import java.util.List;
-
 import io.takari.bpm.IndexedProcessDefinition;
 import io.takari.bpm.ProcessDefinitionUtils;
 import io.takari.bpm.actions.Action;
+import io.takari.bpm.actions.MakeSubProcessVariablesAction;
 import io.takari.bpm.actions.PopCommandAction;
 import io.takari.bpm.actions.PushCommandAction;
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.commands.Command;
 import io.takari.bpm.commands.HandleRaisedErrorCommand;
+import io.takari.bpm.commands.MergeVariablesCommand;
 import io.takari.bpm.commands.ProcessElementCommand;
 import io.takari.bpm.model.ProcessDefinition;
 import io.takari.bpm.model.StartEvent;
+import io.takari.bpm.model.SubProcess;
 import io.takari.bpm.state.ProcessInstance;
+
+import java.util.Collections;
+import java.util.List;
 
 public class SubProcessHandler implements ElementHandler {
 
@@ -21,13 +25,23 @@ public class SubProcessHandler implements ElementHandler {
     public List<Action> handle(ProcessInstance state, ProcessElementCommand cmd, List<Action> actions) throws ExecutionException {
         actions.add(new PopCommandAction());
 
+        IndexedProcessDefinition pd = state.getDefinition(cmd.getDefinitionId());
+        SubProcess s = (SubProcess) ProcessDefinitionUtils.findElement(pd, cmd.getElementId());
+
         // add an error handling command to the stack
         Command errorHandlingCmd = new HandleRaisedErrorCommand(cmd.getDefinitionId(), cmd.getElementId(), cmd.getGroupId(),
                 cmd.isExclusive());
         actions.add(new PushCommandAction(errorHandlingCmd));
 
+        if (s.isUseSeparateContext()) {
+            // set a new variables container (aka child's "ExecutionContext") as our current
+            actions.add(new MakeSubProcessVariablesAction(true));
+
+            // restore the original variables
+            actions.add(new PushCommandAction(new MergeVariablesCommand(state.getVariables(), Collections.emptySet())));
+        }
+
         // find an start event of the subprocess
-        IndexedProcessDefinition pd = state.getDefinition(cmd.getDefinitionId());
         ProcessDefinition sub = ProcessDefinitionUtils.findSubProcess(pd, cmd.getElementId());
         StartEvent start = ProcessDefinitionUtils.findStartEvent(sub);