+++ bpm-engine-impl/src/test/java/io/takari/bpm/AttributesTest.java
@@ -10,7 +10,6 @@ import java.util.Collections;
 import java.util.UUID;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.*;
 
 public class AttributesTest extends AbstractEngineTest {