+++ bpm-engine-impl/src/test/java/io/takari/bpm/CallActivityTest.java
@@ -9,7 +9,6 @@ import org.junit.Test;
 import java.util.*;
 
 import static org.junit.Assert.assertEquals;
-import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.*;
 
 public class CallActivityTest extends AbstractEngineTest {